Fix #32: use the encoding name exposed by DOM and HTML <form>
This would be a mostly editorial change, except that HTML <form> also
exposes the encoding name through its _charset_ feature and does not
have the same mapping table that DOM does. So technically per the
specification to date _charset_ would have to use “utf-8”, not “UTF-8”,
which does not match implementations at all.

With this change only TextDecoder/TextEncoder expose name lowercased as
that was already implemented as such. The encouragement for new places
to expose “utf-8” lowercase also remains.

DOM can now remove its mapping table as it is redundant with this
change.