Note >8835 pointers in index jis0208 cannot be reached
In EUC-JP and ISO-2022-JP encoders, when getting a pointer for a code
point in index jis0208, it’s important to note that the pointer is
always less than 8836. This is because the code points are duplicated
throughout the table and the index pointer returns the first.

Per a suggestion in #47.