Add a BOM sniffing hook for better integration with HTML
This change moves the BOM splitting part of the decode hook into a
separate hook which does not consume any bytes of the token stream.

This will allow fixing a long-standing issue in the HTML encoding
sniffing algorithm with the document's character encoding being set to
the wrong result when there is a BOM (see whatwg/html#1077).

Closes #128.