Specify the usage of token queues for I/O
For a long time, token queues (formerly "streams") had been defined as
simple data structures, with no reference to I/O except for a few uses
of a "wait until" operation not defined anywhere and a comment on the
spec's source saying that the read operation blocks on I/O.

This change introduces the concept of "blocking tokens", which are
allowed to block when being read from the token queue. This makes it
possible for token queues to be used as a direct spec representation of
implementation-internal I/O streams, while all other operations done on
token queues use immediate tokens.

See also #215, which introduced the name "token queues" and defined the
implicit conversions with strings and byte sequences.