Rename streams to I/O queues
This change renames the Encoding-specific concept of "streams", which
had been causing confusion with readable/writable streams for years, to
"I/O queues". It also refactors the I/O queue operations and exports the
corresponding definitions.

As part of this refactoring, "end-of-queue" (formerly "end-of-stream")
becomes an optional item in the queue, indicating that the end of the
streaming data has been reached and that no more data is expected. As a
result, the "read" operation explicitly blocks when trying to read from
an empty queue – a behavior that was previously left unstated.

Closes #180.